+++ weblate/trans/tests/test_hooks.py
@@ -1481,7 +1481,7 @@ class GitHubBackendTest(HookBackendTestCase):
             GITHUB_NEW_PAYLOAD.replace("defunkt/github", "defunkt/git.hub"),
             {
                 "branch": "main",
-                "full_name": "defunkt/github.git",
+                "full_name": "defunkt/github",
                 "repo_url": "http://github.com/defunkt/git.hub",
                 "repos": [
                     "git://github.com/defunkt/git.hub.git",
@@ -1505,7 +1505,7 @@ class BitbucketBackendTest(HookBackendTestCase):
             BITBUCKET_PAYLOAD_GIT,
             {
                 "branch": "main",
-                "full_name": "marcus/project-x.git",
+                "full_name": "marcus/project-x",
                 "repo_url": "https://bitbucket.org/marcus/project-x/",
                 "repos": [
                     "ssh://git@bitbucket.org/marcus/project-x.git",
@@ -1524,7 +1524,7 @@ class BitbucketBackendTest(HookBackendTestCase):
             BITBUCKET_PAYLOAD_HG,
             {
                 "branch": "featureA",
-                "full_name": "marcus/project-x.git",
+                "full_name": "marcus/project-x",
                 "repo_url": "https://bitbucket.org/marcus/project-x/",
                 "repos": [
                     "https://bitbucket.org/marcus/project-x",
@@ -1541,7 +1541,7 @@ class BitbucketBackendTest(HookBackendTestCase):
             BITBUCKET_PAYLOAD_HG_NO_COMMIT,
             {
                 "branch": None,
-                "full_name": "marcus/project-x.git",
+                "full_name": "marcus/project-x",
                 "repo_url": "https://bitbucket.org/marcus/project-x/",
                 "repos": [
                     "https://bitbucket.org/marcus/project-x",
@@ -1558,7 +1558,7 @@ class BitbucketBackendTest(HookBackendTestCase):
             BITBUCKET_PAYLOAD_WEBHOOK,
             {
                 "branch": "name-of-branch",
-                "full_name": "team_name/repo_name.git",
+                "full_name": "team_name/repo_name",
                 "repo_url": "https://api.bitbucket.org/bitbucket/bitbucket",
                 "repos": [
                     "ssh://git@api.bitbucket.org/team_name/repo_name.git",
@@ -1591,7 +1591,7 @@ class BitbucketBackendTest(HookBackendTestCase):
             BITBUCKET_PAYLOAD_HOSTED,
             {
                 "branch": "develop",
-                "full_name": "~DSNOECK/weblate-training.git",
+                "full_name": "~DSNOECK/weblate-training",
                 "repo_url": "https://example.com/weblate-training/browse",
                 "repos": [
                     "ssh://git@bitbucket.org/~DSNOECK/weblate-training.git",
@@ -1648,7 +1648,7 @@ class BitbucketBackendTest(HookBackendTestCase):
                     "ssh://hg@bitbucket.org/team_name/repo_name",
                 ],
                 "branch": "target",
-                "full_name": "team_name/repo_name.git",
+                "full_name": "team_name/repo_name",
             },
         )
 
@@ -1663,7 +1663,7 @@ class BitbucketBackendTest(HookBackendTestCase):
                     "ssh://git@example.com:7999/wlt/locre.git",
                 ],
                 "branch": None,
-                "full_name": "WLT/locre.git",
+                "full_name": "WLT/locre",
             },
         )
 
@@ -1672,7 +1672,7 @@ class BitbucketBackendTest(HookBackendTestCase):
             BITBUCKET_PAYLOAD_WEBHOOK_CLOSED,
             {
                 "branch": "name-of-branch",
-                "full_name": "team_name/repo_name.git",
+                "full_name": "team_name/repo_name",
                 "repo_url": "https://api.bitbucket.org/bitbucket/bitbucket",
                 "repos": [
                     "ssh://git@api.bitbucket.org/team_name/repo_name.git",
@@ -1705,7 +1705,7 @@ class BitbucketBackendTest(HookBackendTestCase):
             BITBUCKET_PAYLOAD_SERVER,
             {
                 "branch": "main",
-                "full_name": "SANDPIT/my-repo.git",
+                "full_name": "SANDPIT/my-repo",
                 "repo_url": "https://example.com/projects/SANDPIT/repos/my-repo/browse",
                 "repos": [
                     "https://example.com/scm/sandpit/my-repo.git",