+++ docs/changes.rst
@@ -7,6 +7,7 @@ Not yet released.
 * Improved analysis on the duplicate language alert.
 * Improved accurancy of the consecutive duplicated words check.
 * Improved scaling of sending many notifications.
+* Improved string state handling for subtitle translation.
 
 `All changes in detail <https://github.com/WeblateOrg/weblate/milestone/87?closed=1>`__.