+++ docs/changes.rst
@@ -14,6 +14,7 @@ Not yet released.
 * Added automatic suggestions management link to management pages.
 * Track comment removal/resolving in history.
 * Fixed restoring project backups with linked components.
+* Fixed captcha entering on unsuccessful registration.
 
 `All changes in detail <https://github.com/WeblateOrg/weblate/milestone/87?closed=1>`__.